requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
flake8==7.1.1
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mccabe==0.7.0
multi-key-dict==2.0.3
packaging==24.1
pbr==6.1.0
pycodestyle==2.12.1
pyflakes==3.2.0
pyngrok==7.2.0
python-dotenv==1.0.1
python-jenkins==1.8.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
urllib3==2.2.3
Werkzeug==3.0.4
